refactor: 코드 가독성 향상을 위해 컴포넌트 및 훅으로 분리
